+++ index.html
@@ -40,7 +40,7 @@
             </div>
         </div>
     </div>
-    <script src="js/login.js"></script>
+    <script src="js/login.mjs"></script>
 </body>
 
 </html>
