+++ js/api/apiLinks.mjs
@@ -1,0 +1,7 @@
+//alle api links skal skrives her og exporteres
+
+export const API_BASE_URL = 'https://api.noroff.dev/api/v1';
+//End-points:
+//Register:  /social/auth/register
+//Login:  /social/auth/login
+//Get All posts: /social/posts
